requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
ascii_colors==0.11.4
asttokens==3.0.1
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.5
av==13.1.0
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.5
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloup==3.0.8
coloredlogs==15.0.1
comm==0.2.3
configparser==7.2.0
cryptography==46.0.3
cyclopts==4.2.4
debugpy==1.8.17
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
fastembed==0.7.3
fastmcp==2.13.1
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
glcontext==3.0.0
google-auth==2.43.0
google-genai==1.51.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
IProgress==0.4
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isosurfaces==0.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.12.0
json_repair==0.54
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==25.7.0
langchain==1.0.7
langchain-core==1.0.5
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.3
langchain-openrouter==0.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lightrag-hku==1.4.9.8
loguru==0.7.3
manim==0.19.0
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.21.2
mdurl==0.1.2
mmh3==5.2.0
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
nano-vectordb==0.0.4.3
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.5
onnxruntime==1.23.2
openai==2.8.0
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pageindex==0.2.1
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pathable==0.4.4
pathvalidate==3.3.1
pexpect==4.9.0
pillow==11.3.0
pipmaster==1.0.10
platformdirs==4.5.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.11
pyglm==2.8.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
pyperclip==1.11.0
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.16.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
scipy==1.16.3
screeninfo==0.8.1
SecretStorage==3.4.1
setuptools==80.9.0
six==1.17.0
skia-pathops==0.9.0
sniffio==1.3.1
soupsieve==2.8
srt==3.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
svgelements==1.9.6
sympy==1.14.0
tavily-python==0.7.13
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
voyageai==0.3.5
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wikipedia==1.4.0
Wikipedia-API==0.8.1
wikipedia-mcp==1.6.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
